ROX-29659: Fix inconsistent names anywhere else
